requirements.txt
kafka
wheel
setuptools
twine
pytest
pytest-runner
boto3
pyyaml

build

